Core.MiniDump: Copy functionality from NUnitExtensions here
Rename the classes to suit the new model.

This is the first commit with code. It contains the project structure, which is designed to support two .NET frameworks, using RjBuild (my personal builder that just calls MsBuild on the solutions and does more lint, than anything else) and Visual Studio. The project is intentially set up as two different solutions to be kept compatible with VS-IDE default solutions. The .NET 4.5 functionality is expected to come later.

Set up the GlobalSuppressions to be that what is used in other personal projects, common supporessions globally for the use with SonarLint 4.14, as well as VS2019 16.4

Tested in HELIOS VS IDE; HELIOS RjProj build.bat; and in the local solution. See docs/nuget.md for information.

Issue: HELIOS-1271