NetVersionDump: First version to get the versions of .NET installed
Get the installed versions of .NET on Windows, and show the runtime that is currently running. There is only very limited support for Mono, we can't know what is installed, but can show what we're running on. Most of this code is taken from my already implemented NetVersion app.

To make errors more obvious, MemoryCrashDataDumpFile and MemoryCrashDumpBlock add enumeration and a dump possibility for debugging blocks that can't be easily tested (keys are dynamic depending on the system).

Issue: HELIOS-1280