CrashExport: Simplify dumpers by having a DumpRow
Most of the time, one needs to initialize the headers and then set properties in the row. But it's not possible to use the Keys property from a Dictionary<string, string>, as there is no well defined order in a dictionary, but the table should have a well defined order. Provide a DumpRow that simplifies this code, and update the IDumpTable to dump rows directly. The DumpTable implements the common code.

Issue: HELIOS-1299