CrashReportApp: Properly multitarget .NET 4.0 and .NET 4.5
When testing on Windows XP, it wouldn't run. The root cause is that the "app.config" contains the SKU .NET 4.5, even if the binary is properly compiled against .NET 4.0. This causes the application to ask to install .NET 4.5 on Windows XP (which doesn't exist). Changing the .config file on the target corrected the error.

But on .NET 4.5 systems (Windows 10) we don't want to have .NET 4.0 in the app.config, then the library "apphelper.dll" will be loaded. This might be OK, but if .NET 4.5 is properly requested in the config - this is not needed.

Issue: HELIOS-1305