CrashDump: Compress using DotNetZip 1.13.5 with .NET 4.0
Use the DotNetZip package to compress the contents of the folder and then delete it when compression is completed.

In this implementation, only .NET 4.0 is disabled. The code for compression is conditionally compiled, so that only on .NET 4 it uses the Ionic.Zip namespace. In .NET 4.5, the project file references `CompressDisabled` which is the same class (only the file name is different). When compression in .NET 4.5 is supported, it will be different.

Issue: HELIOS-1292