XmlCrashDumper: Allow the user to specify their own stylesheet
The App.Config file can now be extended to specify the location of the style sheet that should be used instead of the default, to allow your application to show data that might be in addition to the CrashReporter, or to provide an XSL that has a different style to suit the application better.

The AppConfig file should now be modified to provide the following information:

  <configSections>
    <sectionGroup name="CrashReporter">
      <section name="XmlCrashDumper" type="RJCP.Diagnostics.Config.XmlCrashDumper, RJCP.Diagnostics.CrashReporter"/>
    </sectionGroup>
  </configSections>

  <CrashReporter>
    <XmlCrashDumper>
      <StyleSheet name="RJCP.Diagnostics.CrashReporter, RJCP.Diagnostics.CrashExport.Xml.CrashDump.xsl"/>
    </XmlCrashDumper>
  </CrashReporter>

The style sheet is defined as the <AssemblyName>, <ResourceName>. The example above loads the resource from library (i.e. the built in configuration).

Issue: HELIOS-1323