Crash: Improve documentation and use proper CrashDumpFactory
When analyzing code coverage results, saw that the wrong factory was being used (the XmlCrashDumpFactory was being used directly). The design however calls for the usage of the more generic (wrapper) CrashDumpFactory. Later, if we add more factories, the CrashDumpFactory is the object which will choose the correct factory based on user configuration.