CrashReporter: Treat the file name as a crash dump file name
Mostly documentation changes, that the CreateDump(string) method takes an input which is the path to the crash dump file (if it ends with .zip) or a directory (the dump file name is the same but will have the .zip extension added).

Crash.GetCrashPath(prefix: string) is renamed to Crash.GetCrashDir() for clarity (and it only returns a directory now). Before, ThreadWatchdog got a crash file name, but it was really handled as a directory and was confusing before this change.

Issue: HELIOS-1327