CrashReporter: Making dumping information more robust
An InvalidOperationException can occur as a race condition if a thread dies while it's information is being dumped. Catch the exception, and notify the dumper that the row is invalid (so it won't be dumped).

Make the dumper more robust, so if there's an exception, we don't crash the program (a double fault), and instead just log the dumper issue.

Issue: HELIOS-1339