Fix local solution with CodeQuality in a relative path
When the project NUnitExtensions was renamed to CodeQuality, the solution files were not updated properly. This was introduced with 046565c ("NUnitExtensions: Renamed to RJCP.CodeQuality").

Ensure to check out the repository RJCP.CodeQuality in a folder in parallel to this project called 'nunitextensions' for the project to load properly.

Issue: HELIOS-1337