Crashes: Print exception type for First Chance and Unhandled Exceptions
Often when catching exceptions, a new exception is raised with the inner exception. The message for the new exception may be programmatically the same as the inner exception, making it difficult to differentiate between the two. So print the actual exception type in the log also.

A minor clean up of the output is also done.

Issues: HELIOS-1418