XmlCrashDumper: Sanitize output before writing
Crash dumps would fail with an incomplete XML output when executed under the context of the console software 'Cmder'. This was due to a specific environment variable which contained control characters, such as the 'backspace' key. The XmlWriter would enter an error state and crash dumps could not be performed.

Issue: HELIOS-1415