CrashReporter: Add support for .NET Standard 2.1 (and .NET Core)
Support was already provided for .NET 4.6.1, extending for .NET Standard 2.1 required:
* Basics - extending conditional compilation
* Logging - .NET Core doesn't instantiate a TraceSource on its own. CrashReportApp and Unit Tests must instantiate the TraceSource explicitly with the parameters it wants.

A test case with a logical error was fixed (LogData).

Logging uses the TraceSource, which is standard for .NET Framework, but not for .NET Core. For .NET Core we should be using ILogger<T> which needs further work.

Issue: DOTNET-355