CrashReporter: Use LogSource for diagnostics logging
By using the LogSource instead of the TraceSource, it's easier for .NET Core applications to inject logging as required. By moving over to LogSource, code exposing the TraceSource is no longer needed, as it can be set directly to the LogSource library (LogSource.SetLogSource).

.NET Core needs to instantiate locally itself (see Log.cs in CrashReportApp). The code was cleaned up to make it look easier to implement.

Tested by checking the output of CrashReportApp for all supported frameworks that it works as expected.

Issue: DOTNET-364