OSVersion: Rework to use IsWow64Process2 on newer versions of Windows
By using IsWow64Process2 on newer version of Windows, we can test the native and running process. It is shown to work on Intel and AM64. This code is not tested on ARM or ARM64.

Tested to work correctly on Windows XP 32-bit, as well as Windows 10 21H1.

Issue: DOTNET-383