CrashReporter: Refactor ICrashDataExport
The MemoryTraceListener no longer exposes ICrashDataExport, even if it registers with Crash.Data.Providers. The crash provider is now internal as a detail, it shouldn't be accessible.

By separating the two classes, we provide the basis for a .NET Core implementation that uses the crash provider.

Because the TraceListener registers with the provider, we should also remove the provider when disposing. Test cases were updated accordingly to use the MemoryTraceListener. In user code, this is normally a singleton so is never disposed.

Issue: DOTNET-558