Crash Reporter: Implementing a .NET Core ILogger Provider
.NET Desktop/Framework uses the `TraceListener`. The equivalent for .NET Core is the `ILoggerProvider` and the extensions to register with the Host. Add the .NET Core `ILoggerProvider` and other classes, so that applications can properly initialize logging.

The existing example provided works by using a `TraceListener`, and one can add to the collection to enable writing to multiple sinks, but we can't then (or it's very difficult to) use a JSON application settings file to configure logging in a consistent manner how all other .NET Core logging is done. By providing a `ILoggerProvider` and just calling `.AddMemoryLogger()` allows native .NET logging to be available.

The CrashReportApp now uses an `appsettings.json` file for configuration to log to the console and also to the memory logger to show how your application can do this too.

To ensure that all logs between platforms print UTC time, the log time is now converted to UTC. This will correct timestamps for all.

Issue: DOTNET-558