Core: Capture core dumps with exception information
When loading a core dump within Visual Studio Exception, the exception is now highlighted, not the location where the dump occurred.

Exception information was previously disabled, due to access violation exceptions, usually like:

Unhandled Exception: System.AccessViolationException: Attempted to read or write protected memory. This is often an indication that other memory is corrupt.
   at RJCP.Diagnostics.Native.Win32.DbgHelp.MiniDumpWriteDump(IntPtr hProcess, UInt32 ProcessId, SafeFileHandle hFile, MINIDUMP_TYPE DumpType, MINIDUMP_EXCEPTION_INFORMATION& ExceptionParam, IntPtr UserStreamParam, IntPtr CallackParam)

This is now fixed, by setting the correct "Pack" on the MINIDUMP_EXCEPTION_INFORMATION structure.

On .NET Core 2.1 and earlier the `Marshal.GetExceptionPointers()` wasn't available. As this library is built against .NET Standard 2.1, it still isn't available. We use reflection to get this method if it is being built against .NET Core 3.1 or later.

The "bool" type is removed for safety as it is not blittable, and is replaced with an integer type.

Issue: DOTNET-572