SimplePrioMemoryLog: Don't allow a log group with 0 entries
If a log group was configured to be zero, adding a log would result in trying to dequeue an empty queue which would raise an exception.

Now, if the log is set to zero, it is automatically incremented to one. The code that does the check also ensures that it's not possible to dequeue an empty queue as an extra safeguard.

Issue: DOTNET-568