CrashReporter: Fix or suppress .NET 8.0 warnings
Performance `CA1866: Use StartsWith overload that takes a char` is ignored, as this is only available in .NET Core and later. On .NET 4.6.2 it isn't available and it isn't worth making different code paths for this difference.

Performance `CA1861: Avoid constant arrays as arguments` is ignored in test cases as it isn't required, and the code is more readable without it.

Issue: DOTNET-833