CrashReporter: Add references for OS Compatibility
Fix `CrashData.Providers` to provide information on for the supported platforms. This was identifiable through `CA1416` by adding `[SupporedOSPlatform]` attributes.

The `ThreadDump` provider is supported on Windows and Linux (and FreeBSD), so this means we mark the explicit supported platforms. That in turn really means no change in our behaviour an supported platforms, but makes the code a little more correct.

The `WinVerDump` would theoretically run on all platforms without complaining, but it only makes sense to even execute this on Windows, so applying the attribute to the entire class makes sense here.

Suppress instances of `warning CA1416: This call site is reachable on all platforms`. Methods are given this attribute, and then the diagnostic executes the callchain, so the location of the calls can be properly audited.

Issue: DOTNET-938