CrashReporter: Only delete old files if they're at least five minutes old
If the number of files allowed are very small (e.g. zero), and two crashes occur at the same time, one will try to delete the other while it's being created.

To mitigate this, don't delete the files if they're not yet five minutes old.

This rule doesn't apply if disk space is low.

Issue: DOTNET-969