DotNetZip: Fix CVE-2024-48510
This vulnerability only ever existed on the .NET 4.0 version (maintained for compatibility with older Operating Systems). It has been patched with the fix at https://github.com/mihula/ProDotNetZip/commit/18486ad6d13742a07a6755ef6edf60d7458f1854.

See the file `DotNetZip/README.md` for instructions on how to rebuild DotNetZip. This commit contains the changes needed to patch the vulnerability and integrate the binary into the project. The DotNetZip project is *not* build as part of this project. It must be built separately, and then the binaries are explicitly copied here, removing the reference to the NuGet package. This resolves the vulnerability and the warnings.

Note: When generating the NuGet package, the file `RJCP.DotNetZip.pdb` will be present in both the binary package and the source package. There doesn't seem to be an easy option to have the PDB only in the source package as designed (without writing now a custom nuspec file).

Issue: DOTNET-1022